TST: more method-specific test files (#30453)
