+++ pandas/tests/series/test_analytics.py
@@ -6,65 +6,11 @@ import pytest
 import pandas.util._test_decorators as td
 
 import pandas as pd
-from pandas import Categorical, DataFrame, MultiIndex, Series, Timestamp, isna
+from pandas import DataFrame, MultiIndex, Series
 import pandas.util.testing as tm
 
 
 class TestSeriesAnalytics:
-    def test_argsort(self, datetime_series):
-        self._check_accum_op("argsort", datetime_series, check_dtype=False)
-        argsorted = datetime_series.argsort()
-        assert issubclass(argsorted.dtype.type, np.integer)
-
-        # GH 2967 (introduced bug in 0.11-dev I think)
-        s = Series([Timestamp("201301{i:02d}".format(i=i)) for i in range(1, 6)])
-        assert s.dtype == "datetime64[ns]"
-        shifted = s.shift(-1)
-        assert shifted.dtype == "datetime64[ns]"
-        assert isna(shifted[4])
-
-        result = s.argsort()
-        expected = Series(range(5), dtype="int64")
-        tm.assert_series_equal(result, expected)
-
-        result = shifted.argsort()
-        expected = Series(list(range(4)) + [-1], dtype="int64")
-        tm.assert_series_equal(result, expected)
-
-    def test_argsort_stable(self):
-        s = Series(np.random.randint(0, 100, size=10000))
-        mindexer = s.argsort(kind="mergesort")
-        qindexer = s.argsort()
-
-        mexpected = np.argsort(s.values, kind="mergesort")
-        qexpected = np.argsort(s.values, kind="quicksort")
-
-        tm.assert_series_equal(mindexer, Series(mexpected), check_dtype=False)
-        tm.assert_series_equal(qindexer, Series(qexpected), check_dtype=False)
-        msg = (
-            r"ndarray Expected type <class 'numpy\.ndarray'>,"
-            r" found <class 'pandas\.core\.series\.Series'> instead"
-        )
-        with pytest.raises(AssertionError, match=msg):
-            tm.assert_numpy_array_equal(qindexer, mindexer)
-
-    def _check_accum_op(self, name, datetime_series_, check_dtype=True):
-        func = getattr(np, name)
-        tm.assert_numpy_array_equal(
-            func(datetime_series_).values,
-            func(np.array(datetime_series_)),
-            check_dtype=check_dtype,
-        )
-
-        # with missing values
-        ts = datetime_series_.copy()
-        ts[::2] = np.NaN
-
-        result = func(ts)[1::2]
-        expected = func(np.array(ts.dropna()))
-
-        tm.assert_numpy_array_equal(result.values, expected, check_dtype=False)
-
     def test_compress(self):
         cond = [True, False, True, False, False]
         s = Series([1, -1, 5, 8, 7], index=list("abcde"), name="foo")
@@ -363,100 +309,3 @@ class TestSeriesAnalytics:
         )
         with pytest.raises(ValueError, match=msg):
             np.sum(s, keepdims=True)
-
-
-class TestCategoricalSeriesAnalytics:
-    @pytest.mark.parametrize(
-        "dtype",
-        ["int_", "uint", "float_", "unicode_", "timedelta64[h]", "datetime64[D]"],
-    )
-    def test_drop_duplicates_categorical_non_bool(self, dtype, ordered_fixture):
-        cat_array = np.array([1, 2, 3, 4, 5], dtype=np.dtype(dtype))
-
-        # Test case 1
-        input1 = np.array([1, 2, 3, 3], dtype=np.dtype(dtype))
-        tc1 = Series(Categorical(input1, categories=cat_array, ordered=ordered_fixture))
-        if dtype == "datetime64[D]":
-            # pre-empty flaky xfail, tc1 values are seemingly-random
-            if not (np.array(tc1) == input1).all():
-                pytest.xfail(reason="GH#7996")
-
-        expected = Series([False, False, False, True])
-        tm.assert_series_equal(tc1.duplicated(), expected)
-        tm.assert_series_equal(tc1.drop_duplicates(), tc1[~expected])
-        sc = tc1.copy()
-        sc.drop_duplicates(inplace=True)
-        tm.assert_series_equal(sc, tc1[~expected])
-
-        expected = Series([False, False, True, False])
-        tm.assert_series_equal(tc1.duplicated(keep="last"), expected)
-        tm.assert_series_equal(tc1.drop_duplicates(keep="last"), tc1[~expected])
-        sc = tc1.copy()
-        sc.drop_duplicates(keep="last", inplace=True)
-        tm.assert_series_equal(sc, tc1[~expected])
-
-        expected = Series([False, False, True, True])
-        tm.assert_series_equal(tc1.duplicated(keep=False), expected)
-        tm.assert_series_equal(tc1.drop_duplicates(keep=False), tc1[~expected])
-        sc = tc1.copy()
-        sc.drop_duplicates(keep=False, inplace=True)
-        tm.assert_series_equal(sc, tc1[~expected])
-
-        # Test case 2
-        input2 = np.array([1, 2, 3, 5, 3, 2, 4], dtype=np.dtype(dtype))
-        tc2 = Series(Categorical(input2, categories=cat_array, ordered=ordered_fixture))
-        if dtype == "datetime64[D]":
-            # pre-empty flaky xfail, tc2 values are seemingly-random
-            if not (np.array(tc2) == input2).all():
-                pytest.xfail(reason="GH#7996")
-
-        expected = Series([False, False, False, False, True, True, False])
-        tm.assert_series_equal(tc2.duplicated(), expected)
-        tm.assert_series_equal(tc2.drop_duplicates(), tc2[~expected])
-        sc = tc2.copy()
-        sc.drop_duplicates(inplace=True)
-        tm.assert_series_equal(sc, tc2[~expected])
-
-        expected = Series([False, True, True, False, False, False, False])
-        tm.assert_series_equal(tc2.duplicated(keep="last"), expected)
-        tm.assert_series_equal(tc2.drop_duplicates(keep="last"), tc2[~expected])
-        sc = tc2.copy()
-        sc.drop_duplicates(keep="last", inplace=True)
-        tm.assert_series_equal(sc, tc2[~expected])
-
-        expected = Series([False, True, True, False, True, True, False])
-        tm.assert_series_equal(tc2.duplicated(keep=False), expected)
-        tm.assert_series_equal(tc2.drop_duplicates(keep=False), tc2[~expected])
-        sc = tc2.copy()
-        sc.drop_duplicates(keep=False, inplace=True)
-        tm.assert_series_equal(sc, tc2[~expected])
-
-    def test_drop_duplicates_categorical_bool(self, ordered_fixture):
-        tc = Series(
-            Categorical(
-                [True, False, True, False],
-                categories=[True, False],
-                ordered=ordered_fixture,
-            )
-        )
-
-        expected = Series([False, False, True, True])
-        tm.assert_series_equal(tc.duplicated(), expected)
-        tm.assert_series_equal(tc.drop_duplicates(), tc[~expected])
-        sc = tc.copy()
-        sc.drop_duplicates(inplace=True)
-        tm.assert_series_equal(sc, tc[~expected])
-
-        expected = Series([True, True, False, False])
-        tm.assert_series_equal(tc.duplicated(keep="last"), expected)
-        tm.assert_series_equal(tc.drop_duplicates(keep="last"), tc[~expected])
-        sc = tc.copy()
-        sc.drop_duplicates(keep="last", inplace=True)
-        tm.assert_series_equal(sc, tc[~expected])
-
-        expected = Series([True, True, True, True])
-        tm.assert_series_equal(tc.duplicated(keep=False), expected)
-        tm.assert_series_equal(tc.drop_duplicates(keep=False), tc[~expected])
-        sc = tc.copy()
-        sc.drop_duplicates(keep=False, inplace=True)
-        tm.assert_series_equal(sc, tc[~expected])
